+++ sources/mandatory/algebra/tuple.c
@@ -6,18 +6,18 @@
 /*   By: lpaulo-m <lpaulo-m@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/26 15:11:13 by lpaulo-m          #+#    #+#             */
-/*   Updated: 2022/12/26 19:07:20 by lpaulo-m         ###   ########.fr       */
+/*   Updated: 2022/12/26 19:19:37 by lpaulo-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minirt.h>
 
-t_t3d	tuple(double x, double y, double z, int type)
+t_t3d	tuple(double x, double y, double z, double type)
 {
 	return (tuple_3d(x, y, z, type));
 }
 
-t_t3d	tuple_3d(double x, double y, double z, int type)
+t_t3d	tuple_3d(double x, double y, double z, double type)
 {
 	return ((t_t3d){x, y, z, type});
 }
@@ -29,5 +29,5 @@ t_t3d	null_tuple(void)
 
 void	inspect_tuple(t_t3d tuple)
 {
-	printf("TUPLE: %f %f %f, type %d\n", tuple.x, tuple.y, tuple.z, tuple.type);
+	printf("TUPLE: %f %f %f, type %f\n", tuple.x, tuple.y, tuple.z, tuple.type);
 }
